+++ internal/generator/utils.go
@@ -1,13 +1,71 @@
 package generator
 
 import (
+	"bytes"
 	"fmt"
 	"go/ast"
 	"strconv"
+	"strings"
 	"time"
 	"unicode"
 	"unicode/utf8"
 )
+
+func prefix(str1, str2 string) string {
+	if str1 == "" || str2 == "" {
+		return ""
+	}
+
+	buf := bytes.NewBuffer(nil)
+	if len(str1) > len(str2) {
+		str1, str2 = str2, str1
+	}
+
+	for i := range str1 {
+		if str1[i] != str2[i] {
+			break
+		}
+
+		buf.WriteByte(str1[i])
+	}
+
+	return buf.String()
+}
+
+func formatComment(comment string) string {
+	if comment == "" {
+		return ""
+	}
+
+	buf := bytes.NewBuffer(nil)
+
+	lines := strings.Split(comment, "\n")
+	// this is a hack for go1.18.
+	//nolint:godox
+	// TODO: need help with go build constraints. failed to configure tags like // go:build go1.18
+	commonPrefix := lines[0]
+	for _, line := range lines {
+		commonPrefix = prefix(commonPrefix, line)
+	}
+
+	for i := range lines {
+		lines[i] = strings.TrimPrefix(lines[i], commonPrefix)
+
+		// Last line contains an empty string.
+		if lines[i] == "" && i == len(lines)-1 {
+			continue
+		}
+
+		if i != 0 {
+			buf.WriteString("\n")
+		}
+
+		buf.WriteString("// ")
+		buf.WriteString(lines[i])
+	}
+
+	return buf.String()
+}
 
 func findStructTypeParamsAndFields(packages map[string]*ast.Package, typeName string) ([]*ast.Field, []*ast.Field) {
 	decls := getDecls(packages)
